+++ backend-rust/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add `SearchResult::bakers` and `SearchResult::contracts`.
 - Add query `Query::poolRewardMetricsForPassiveDelegation`.
 - Add REST API `/rest/export/statement` for exporting account statements as CSV.